:lipstick: Move category stats to a dedicated page
