+++ noplp/song.py
@@ -3,6 +3,7 @@
 from dataclasses import dataclass
 from datetime import date
 
+
 @dataclass
 class Song:
     """Class used to manage song instances imported from API data.
@@ -13,6 +14,7 @@ class Song:
         title: A string with the title of the song.
         lyrics: An string with the lyrics obtained from the Wiki API.
     """
+
     title: str
     singer: str
     lyrics: str
@@ -20,7 +22,6 @@ class Song:
     categories: list[str]
     points: list[int]
     emissions: list[int]
-
 
     def __str__(self) -> str:
         return f"Song instance of the title '{self.title}'"
